SilentBoss v1.5.0
Fixed an issue affecting the client where if the wither was spawned the
SilenceEnderDragon config option was checked instead of SilenceWither
option.

Added ability to change config options via the following:
- Main Menu: Mods -> SilentBoss -> Config
- In Game: Esc -> Mod Options... -> SilentBoss -> Config

Improvements to server side packet handling, it now finds the
SPacketEffect id instead of assuming it.

Updated to Forge 1.10.2-12.18.3.2488

Supports 1.10.2 - 1.12.2